Fix bad type conversion in call to trl_set_locking. Change "inline" to "static inline" to avoid compile error.

+++ INC/trlcore_i.h
@@ -564,7 +564,7 @@ void trl_get_eval(int nd, int locked, double *alpha, double *beta,
 ////
 */
 void trl_set_locking(int jnd, int nlam, double *lambda, double *res,
-		      double *yy, int anrm, int *locked);
+		      double *yy, double anrm, int *locked);
 /*
 // Purpose
 // =======
